% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpter_run.R
\name{gpter_run}
\alias{gpter_run}
\title{Send a prompt to the ChatGPT API and get a response}
\usage{
gpter_run(prompt, stop_sequence = NULL)
}
\arguments{
\item{prompt}{The text prompt to send to the ChatGPT API}

\item{stop_sequence}{null value(No use)}
}
\value{
the result of the code
}
\description{
Send a prompt to the ChatGPT API and get a response
}
\examples{
\dontrun{
  iniSet("your_api_key")
  prompt <- "What is the capital of France?"
  gpter_run(prompt)
}

}
