% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iniSet.R
\name{iniSet}
\alias{iniSet}
\title{Set the ChatGPT API key for the package}
\usage{
iniSet(api_key)
}
\arguments{
\item{api_key}{Your ChatGPT API key}
}
\description{
Set the ChatGPT API key for the package
}
\examples{
\dontrun{
  api_key <- "your_api_key"
  iniSet(api_key)
}
}
